Refactor: Resolve circular dependencies
This commit resolves several circular dependencies identified in the codebase:

1.  **prompt/build.mjs and its consumers:**
    *   Moved `margePrompt` from `prompt/build.mjs` to a new `prompt/utils.mjs` file.
    *   Updated `prompt/build.mjs`, `prompt/ads/index.mjs`, `prompt/functions/index.mjs`, `prompt/memory/index.mjs`, `prompt/role_settings/index.mjs`, and `prompt/system/index.mjs` to import `margePrompt` from the new utility file.

2.  **scripts/match.mjs and scripts/langdetect.mjs:**
    *   Moved `base_match_keys` and its helper functions (`SimplifiyChinese`, `SimpleSimplify`) from `scripts/match.mjs` to a new `scripts/match_utils.mjs` file.
    *   Updated `scripts/match.mjs` and `scripts/langdetect.mjs` to import these functions from the new utility file, eliminating code duplication and resolving the circular dependency.

3.  **reply_gener/index.mjs and its function modules:**
    *   Moved `GetReply` and `getLongTimeLogAdder` from `reply_gener/index.mjs` to a new `reply_gener/core.mjs` file.
    *   Updated `reply_gener/index.mjs` to import and re-export these functions from `reply_gener/core.mjs`.
    *   Updated `reply_gener/functions/coderunner.mjs`, `reply_gener/functions/detail-thinking.mjs`, and `reply_gener/functions/timer.mjs` to import the moved functions directly from `reply_gener/core.mjs`.

These changes break the import cycles, leading to a more stable and maintainable module structure. A syntax check was performed on all .mjs files and passed successfully.
