+++ prompt/ads/index.mjs
@@ -1,4 +1,4 @@
-import { margePrompt } from '../build.mjs'
+import { margePrompt } from '../utils.mjs'
 import { ps12exePrompt } from './ps12exe.mjs'
 /** @typedef {import("../../../../../../../src/public/shells/chat/decl/chatLog.ts").chatReplyRequest_t} chatReplyRequest_t */
 /** @typedef {import("../logical_results/index.mjs").logical_results_t} logical_results_t */
