+++ reply_gener/functions/coderunner.mjs
@@ -2,7 +2,7 @@ import process from 'node:process'
 import util from 'node:util'
 import { bash_exec, pwsh_exec } from '../../scripts/exec.mjs'
 import { async_eval } from '../../scripts/async_eval.mjs'
-import { GetReply } from '../index.mjs'
+import { GetReply } from '../core.mjs'
 import { newCharReplay, statisticDatas } from '../../scripts/statistics.mjs'
 import { captureScreen } from '../../scripts/tools.mjs'
 import { toFileObj } from '../../scripts/fileobj.mjs'
