+++ reply_gener/functions/detail-thinking.mjs
@@ -10,7 +10,7 @@ import { webbrowse } from './webbrowse.mjs'
 import { googlesearch } from './googlesearch.mjs'
 import { coderunner } from './coderunner.mjs'
 import { GoogleSearchPrompt } from '../../prompt/functions/googlesearch.mjs'
-import { getLongTimeLogAdder } from '../index.mjs'
+import { getLongTimeLogAdder } from '../core.mjs'
 import { config } from '../../config.mjs'
 import { sleep } from '../../scripts/tools.mjs'
 import { statisticDatas } from '../../scripts/statistics.mjs'
